Some Dutch translations fixes and added keys
The main translation file was missing a bunch of keys, mostly for the menu headers. Added those. Yay for autism. Please add the missing keys (bottom of the file) to the main file as well. In its current state this file does work in the menu (tested it) so you are able to merge it.